refactor(test): one countingStore helper instead of three near-copies
The store-call counting Proxy existed three times: byte-identical in the two
round-trip budget tests, and a third variant in authz-memoization that counted
a fixed set of methods into a keyed record. All three carried their own copy of
the same hard-won comment about why it must be a `get`-level Proxy rather than
patched methods (the pg test store is itself a Proxy, so patching counts
nothing — a mistake that once made a test pass on SQLite while measuring
nothing at all on Postgres).

One `countingStore` in helpers.ts now backs all three, with that explanation in
a single place. The keyed variant is just `countOf(name)` over the recorded
calls, so authz-memoization's assertions read the same and check the same
thing. `reset()` replaces the hand-rolled `calls.length = 0`.
