+++ apps/api/test/authz-memoization.test.ts
@@ -1,6 +1,6 @@
 import type { MetaStore } from "@derive/core"
 import { describe, expect, it } from "vitest"
-import { as, jsonAs, makeAuthedApp, publishAs, type TestUser } from "./helpers"
+import { as, countingStore, jsonAs, makeAuthedApp, publishAs, type TestUser } from "./helpers"
 
 /**
  * ONE ACTOR RESOLUTION PER REQUEST, PER ARTIFACT.
@@ -15,10 +15,8 @@ import { as, jsonAs, makeAuthedApp, publishAs, type TestUser } from "./helpers"
  * with the same membership row fetched three times. On the hosted edge each is ~100-900ms,
  * because the isolate and Postgres sit in different regions.
  *
- * Asserted at the STORE BOUNDARY, where the cost is, and counted through a `get`-level Proxy
- * rather than by patching methods — the pg test store is itself a Proxy deferring to an
- * async-created store, so assigning over its methods silently counts nothing. That is how the
- * first version of this test passed on SQLite while measuring nothing at all on Postgres.
+ * Asserted at the STORE BOUNDARY, where the cost is, via the shared `countingStore` helper
+ * (see helpers.ts for why it counts through a Proxy rather than by patching methods).
  */
 const owner: TestUser = { id: "u_memo_own", email: "memo@derive.test", name: "Owner" }
 
@@ -28,27 +26,6 @@ const COUNTED = [
   "collectionRolesForArtifact",
   "artifactGrants",
 ] as const
-type Counted = (typeof COUNTED)[number]
-
-const counting = (inner: MetaStore) => {
-  const counts: Record<Counted, number> = {
-    getMembership: 0,
-    getArtifactMember: 0,
-    collectionRolesForArtifact: 0,
-    artifactGrants: 0,
-  }
-  const proxy = new Proxy(inner, {
-    get(target, prop, recv) {
-      const value = Reflect.get(target, prop, recv)
-      if (!COUNTED.includes(prop as Counted) || typeof value !== "function") return value
-      return (...args: unknown[]) => {
-        counts[prop as Counted] += 1
-        return (value as (...a: unknown[]) => unknown).apply(target, args)
-      }
-    },
-  })
-  return { proxy, counts }
-}
 
 // The first app owns the store; the second drives the request through a counting wrapper around
 // that same store, so both see identical data.
@@ -58,7 +35,7 @@ const counting = (inner: MetaStore) => {
 // pg_namespace_nspname_index". Its own store is then built and never used, because `deps.meta`
 // overrides it with the wrapper below; only the schema name has to differ.
 const base = makeAuthedApp("authz-memo", [owner])
-const { proxy, counts } = counting(base.meta as MetaStore)
+const { proxy, countOf, reset } = countingStore(base.meta as MetaStore)
 const { app } = makeAuthedApp("authz-memo-probe", [owner], undefined, { deps: { meta: proxy } })
 
 describe("actor resolution is memoized per request", () => {
@@ -71,7 +48,7 @@ describe("actor resolution is memoized per request", () => {
     const published = await publishAs(base.app, "# Memo\n\n## One\n\nBody.\n", {}, as(owner.email))
     const { short_id } = (await published.json()) as { short_id: string }
 
-    for (const k of COUNTED) counts[k] = 0
+    reset()
     const res = await app.request(
       "/v1/artifacts/chat-session",
       jsonAs(as(owner.email), { short_id, body_md: "How many sections?", mode: "publish" }),
@@ -82,20 +59,20 @@ describe("actor resolution is memoized per request", () => {
 
     // The store WAS consulted. A zero here means the wrapper missed, not that the cache worked —
     // which is exactly how this test previously lied on Postgres.
-    const total = COUNTED.reduce((n, k) => n + counts[k], 0)
+    const total = COUNTED.reduce((n, k) => n + countOf(k), 0)
     expect(total, "counting proxy saw no permission reads at all").toBeGreaterThan(0)
 
-    if (counts.artifactGrants > 0) {
+    if (countOf("artifactGrants") > 0) {
       // Fast path: one call answers everything, and the reads it replaces stay untouched.
-      expect(counts.artifactGrants).toBe(1)
-      expect(counts.getMembership).toBe(0)
-      expect(counts.getArtifactMember).toBe(0)
-      expect(counts.collectionRolesForArtifact).toBe(0)
+      expect(countOf("artifactGrants")).toBe(1)
+      expect(countOf("getMembership")).toBe(0)
+      expect(countOf("getArtifactMember")).toBe(0)
+      expect(countOf("collectionRolesForArtifact")).toBe(0)
     } else {
       // Fallback path: each permission row read exactly once for the whole request.
-      expect(counts.getMembership).toBe(1)
-      expect(counts.getArtifactMember).toBe(1)
-      expect(counts.collectionRolesForArtifact).toBe(1)
+      expect(countOf("getMembership")).toBe(1)
+      expect(countOf("getArtifactMember")).toBe(1)
+      expect(countOf("collectionRolesForArtifact")).toBe(1)
     }
   })
 })
