+++ apps/api/test/helpers.ts
@@ -468,3 +468,44 @@ export const pub = (
   const h = Object.keys(headers).length ? headers : TOKEN_HEADER
   return app.request(url, { method: "POST", body: form, headers: h })
 }
+
+/**
+ * Wrap a store so every method call is recorded by name, in order.
+ *
+ * The round-trip budgets and the authorization-memoization test both assert at the STORE
+ * BOUNDARY, because that is where the ~80ms edge cost is (see round-trip-budget.test.ts).
+ *
+ * Counting happens in a `get`-level Proxy rather than by patching methods, and that detail
+ * is load-bearing: the pg test store is ITSELF a Proxy deferring to an async-created store,
+ * so assigning over its methods silently counts nothing. That mistake once made a test pass
+ * on SQLite while measuring absolutely nothing on Postgres — which is the whole reason this
+ * lives in one place now instead of being re-derived per test file.
+ *
+ * Pair it with a second `makeAuthedApp` whose `deps.meta` is the returned proxy: the first
+ * app owns the store, the probe app drives requests through the wrapper around that same
+ * store, so both see identical data. Give the probe its own NAME — two apps sharing a name
+ * share a Postgres schema and race to create it.
+ */
+export const countingStore = (inner: MetaStore) => {
+  const calls: string[] = []
+  const proxy = new Proxy(inner, {
+    get(target, prop, recv) {
+      const value = Reflect.get(target, prop, recv)
+      if (typeof value !== "function" || typeof prop !== "string") return value
+      return (...args: unknown[]) => {
+        calls.push(prop)
+        return (value as (...a: unknown[]) => unknown).apply(target, args)
+      }
+    },
+  })
+  return {
+    proxy,
+    /** Every call made since the last `reset()`, in order — the budget unit. */
+    calls,
+    reset: () => {
+      calls.length = 0
+    },
+    /** How many times one method was called. */
+    countOf: (method: string) => calls.filter((c) => c === method).length,
+  }
+}
